Program of smart bike data recorder for accident recovery.
INTRODUCTION
              Now a day in every country vehicle accident is a major problem. Bikes are more likely involved in overall crashes. Many services
are conducted by the different campaign for the awareness of death and disabilities due to the accident. This result in the
development of a safety system also different technologies are invented to overcome this problem Here we are making efforts to
overcome bike accidents. Many of the accident takes place due to the negligence of driver and breaking traffic rules. Drunken
driving is one of the major issues also not wearing a helmet while riding a bike [1]. In the existing system, a GPS tracking device
is implemented for the real-time monitoring system. It will be helpful when an accident occurs by giving the location in terms of
latitude and longitude to the victim’s family and nearby ambulance via GSM (Global System for mobile) [2]. Various sensors like
Temperature sensor, Speed sensor, Helmet switch, etc. are used for the safety of drive [3]. These sensors will give information
about the different parameters of the bike.
In airplanes black box traditionally being used for investigating an airplane accident. Now a day this is not the key-feature only for
the airplanes in high consumer vehicle black box like the system is used [4]. But not in bikes because of cost issues. In the system,
we are implementing black box like recorder called smart recorder to record real-time speed, tilt, the status of eyes of rider i.e.
open or close using image processing technology [5]. This data is also stored on remote IOT server. This stored data is useful for
police investigation purpose hence finding the real culprit. We can see also the real scenario. The insurance company uses this
recorded data for checking the policy terms also for calculating insurance by monitoring data regularly. 


SYSTEM OVERVIEW
         This smart bike recorder has different sensors interfaced to the raspberry pi module. These sensors are a temperature sensor, an
alcohol sensor, accelerometer, tilt sensor. After wearing the helmet, helmet switch checks it and grant permission for starting the
bike. Alcohol sensor senses the presence of alcohol if not found in fact then and only then the bike will close the ignition relay and
bike will start. Temperature sensor sense amount of the engine temperature exhausted from the bike. This helps in maintaining the
engine of the bike. The accelerometer measures the speed of the bike and warns the rider if he crosses the speed limit. This speed
is regularly stored on IOT server in graphical form. Tilt sensor measures the tilting position with reference to the bike. They
enable the easy detection of the orientation of bike. The webcam takes images at certain intervals of the eyes of the rider and using
image processing, we monitor the position of eyes. Using this we can check the rider is sleeping or not. GPS is used for real-time
tracking of the bike. All the information about the sensor is given to the microcontroller (here raspberry-pi) and through it to the
IOT server.
The occurrence of an accident is detected by calculating parameters stored in the IOT server. After an incident, the message is sent
to the family member and nearby ambulance using GSM and also the position of the bike is also sent in the form of latitude and
longitude